applications: asset_tracker_v2: Cleanup P-GPS integration
Moved P-GPS request handling from data module to cloud
module. Simplified P-GPS request handling and aligned
P-GPS and A-GNSS implementations.

Added a missing value to cloud module backoff delay table.

Did a couple of minor documentation fixes.

NCSDK-18093
