+++ applications/asset_tracker_v2/doc/cloud_module.rst
@@ -82,7 +82,7 @@ nRF Cloud A-GNSS and P-GPS
 ==========================
 
 When the cloud module is configured to communicate with `AWS IoT Core`_, `Azure IoT Hub`_, or an `LwM2M`_ server, it supports processing of received A-GNSS and P-GPS data using the :ref:`lib_nrf_cloud_agnss` and :ref:`lib_nrf_cloud_pgps` libraries.
-This enables the cloud service to fetch A-GNSS and P-GPS data directly from `nRF Cloud`_ using REST calls and relay this data to an nRF91 Series SiP using the pre-established cloud connection.
+This enables the cloud service to fetch A-GNSS and P-GPS data directly from `nRF Cloud`_ and relay this data to an nRF91 Series SiP using the pre-established cloud connection.
 By reusing the pre-established connection, the application saves overhead related to maintaining multiple connections at the same time.
 When configuring the application to communicate with nRF Cloud, A-GNSS and P-GPS data are received directly from the service, and not by proxy.
 For more information, see `nRF Cloud Location Services <nRF Cloud Location Services documentation_>`_.
@@ -114,7 +114,7 @@ Reconnection is implemented with a binary backoff based on the following lookup 
 .. code-block:: c
 
    static struct cloud_backoff_delay_lookup backoff_delay[] = {
-      { 32 }, { 64 }, { 128 }, { 256 }, { 512 },
+      { 32 }, { 64 }, { 128 }, { 256 }, { 512 }, { 1024 },
       { 2048 }, { 4096 }, { 8192 }, { 16384 }, { 32768 },
       { 65536 }, { 131072 }, { 262144 }, { 524288 }, { 1048576 }
    };
