+++ applications/asset_tracker_v2/src/events/data_module_event.c
@@ -28,8 +28,6 @@ static char *get_evt_type_str(enum data_module_event_type type)
 		return "DATA_EVT_IMPACT_DATA_SEND";
 	case DATA_EVT_CLOUD_LOCATION_DATA_SEND:
 		return "DATA_EVT_CLOUD_LOCATION_DATA_SEND";
-	case DATA_EVT_AGNSS_REQUEST_DATA_SEND:
-		return "DATA_EVT_AGNSS_REQUEST_DATA_SEND";
 	case DATA_EVT_CONFIG_INIT:
 		return "DATA_EVT_CONFIG_INIT";
 	case DATA_EVT_CONFIG_READY:
